python/test/requirements.txt
firebase-admin
pandas
plotly
beautifulsoup4
selenium
webdriver-manager
pdfplumber
openai
schedule
xlrd
python-dotenv
openpyxl
PyPDF2
pdfminer.six
